% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vldaft.R
\name{parse_vldaft_formula}
\alias{parse_vldaft_formula}
\title{Parse vldaft formula}
\usage{
parse_vldaft_formula(formula, data)
}
\arguments{
\item{formula}{The formula.}

\item{data}{The data frame.}
}
\value{
A list with response, loc_formula, scale_formula.
}
\description{
Splits a formula like Surv(time, status) ~ x1 + x2 | x3 + x4
into location and scale parts.
}
\keyword{internal}
